Adicionado versionamento da api / conlfigurado automapper / ajustado injeção de dependencia

+++ src/MySales.Core/Entities/Movements.cs
@@ -4,10 +4,11 @@ namespace MySales.Core.Entities;
 
 public class Movements
 {
-    public int Id { get; set; }
-    public int ProductId { get; set; }
+    public long Id { get; set; }
+    public long ProductId { get; set; }
     public TypeMovement Type { get; set; }
     public int Amount { get; set; }
+
     public DateTime CreatedDate { get; set; } = DateTime.Now;
 
     public virtual Product Product { get; set; } = default!;
