+++ src/MySales.Core/Entities/Stock.cs
@@ -2,8 +2,8 @@
 
 public class Stock
 {
-    public int Id { get; set; }
-    public int ProductId { get; set; }
+    public long Id { get; set; }
+    public long ProductId { get; set; }
     public int Amount { get; set; }
 
     public virtual Product Product { get; set; } = default!;
